Day 10 - Solution with help
Day 10 with solution for part 2. I have not validated it on the site because even if I was close I did not manage to implement the right algorithm (I was missing a list to store combinations, among other things).
I was happy to have foud it was like looking for all paths in a DAG and the algorithm to apply but I was not able to implement it. Will try later with dynamic programming.